Artifacts/project/sovereignshield/requirements.txt
# SovereignShield — HuggingFace Space / local
shiny>=1.0.0
anthropic>=0.39.0
supabase>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
